% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anthracite_code.R
\name{anthracite_code}
\alias{anthracite_code}
\title{Form the digital code for anthracite in accordance with GOST 25543}
\usage{
anthracite_code(r, sok, vnudaf, ar)
}
\arguments{
\item{r}{reflectance of vitrinite, [\emph{\%}], measured in accordance with
\strong{ISO 7404-5}.
Type: [\code{double}].}

\item{sok}{total volume of all \emph{fusinite} fractions, [\emph{\%}].
The next \emph{fusinite} fractions may be considered:
\describe{
  \item{\emph{inertinite}}{determined in accordance with \strong{ISO 7404-3}}
  \item{two thirds of \emph{semivitrinite}}{determined in accordance with \href{http://docs.cntd.ru/document/1200105478}{GOST R 55662}}
}
Type: [\code{numeric}].}

\item{vnudaf}{\emph{volatile matter volume yield}, [\emph{cm^3/g}], measured in
accordance with \href{http://docs.cntd.ru/document/1200024047}{GOST 7303}.
Type: [\code{double}].}

\item{ar}{anisotropy of vitrinite reflectance, [\emph{\%}], calculated in accordance
with formula 4 in \href{http://docs.cntd.ru/document/1200105476}{GOST R 55659}.
Type: [\code{double}].}
}
\value{
Digital code according to \emph{Part 7} of
  \href{http://docs.cntd.ru/document/1200107843}{GOST 25543}.
  Type: [\code{character}].
}
\description{
In accordance with
\href{http://docs.cntd.ru/document/1200107843}{GOST 25543} (\emph{part 7})
form the digital code for \emph{anthracite} using results of laboratory
measurements.
}
\examples{
# Consider two samples of anthracite with the next laboratory
# measurement results:
r0   <- c(3.51, 3.64)
sok  <- c(12, 2)
vnudaf <- c(270.3, 174.6)
ar <- c(0, 46.8)

x <- anthracite_code(r0, sok, vnudaf, ar)
print(x)
# [1] "3512020" "3601540"

# Unit test:
stopifnot(
  x \%in\% with(coal.state:::g25543db()$anthracite, {
    c(
      S1ABx$code, S2ABx$code
    )
  })
)
}
