% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anthracite_subtype.R
\name{anthracite_subtype}
\alias{anthracite_subtype}
\title{Determine the subtype of anthracite in accordance with GOST 25543}
\usage{
anthracite_subtype(ar)
}
\arguments{
\item{ar}{anisotropy of vitrinite reflectance, [\emph{\%}], calculated in accordance
with formula 4 in \href{http://docs.cntd.ru/document/1200105476}{GOST R 55659}.
Type: [\code{double}].}
}
\value{
Identifier of anthracite subtype according to \emph{Table 10} of
  \href{http://docs.cntd.ru/document/1200107843}{GOST 25543}.
  Type: [\code{character}].
}
\description{
In accordance with
\href{http://docs.cntd.ru/document/1200107843}{GOST 25543} (\emph{Table 10})
determine the subtype of anthracite by anisotropy of vitrinite reflectance,
[\emph{\%}].
}
\examples{
anthracite_subtype(c(0, 29.4, 29.5, 40, 40.509, 80))
# [1] "20" "20" "30" "30" "40" "70"

# build test:
stopifnot(
  all(
    anthracite_subtype(c(0, 29.4, 29.5, 40, 40.509, 80)) ==
    c("20", "20", "30", "30", "40", "70")
  )
)

}
