% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brown_subtype.R
\name{brown_subtype}
\alias{brown_subtype}
\title{Determine the subtype of brown coal in accordance with GOST 25543}
\usage{
brown_subtype(tscdaf)
}
\arguments{
\item{tscdaf}{\emph{yield of tar}, [\emph{\%}], measured in accordance with
\strong{ISO 647} and recalculated on dry ash-free basis.
Type: [\code{double}].}
}
\value{
Identifier of brown coal subtype according to \emph{Table 8} of
  \href{http://docs.cntd.ru/document/1200107843}{GOST 25543}.
  Type: [\code{character}].
}
\description{
In accordance with
\href{http://docs.cntd.ru/document/1200107843}{GOST 25543} (\emph{Table 8})
determine the subtype of brown coal by \emph{yield of tar}, [\emph{\%}].
}
\examples{
brown_subtype(c(0.1, 9.9, 10.04, 10.05, 15.0, 15.09, 20.0, 43.2))
# [1] "05" "05" "05" "10" "10" "15" "15" "20"

# building test:
stopifnot(
  all(
    brown_subtype(c(0.1, 9.9, 10.04, 10.05, 15.0, 15.09, 20.0, 43.2)) ==
      c("05", "05", "05", "10", "10", "15", "15", "20")
  )
)

}
