% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brown_type.R
\name{brown_type}
\alias{brown_type}
\title{Determine the type of brown coal in accordance with GOST 25543}
\usage{
brown_type(wmaxaf)
}
\arguments{
\item{wmaxaf}{maximum \emph{moisture-holding capacity}, [\emph{\%}], measured in
accordance with \strong{ISO 1018}. Type: [\code{double}].}
}
\value{
Identifier of brown coal type according to \emph{Table 5} of
  \href{http://docs.cntd.ru/document/1200107843}{GOST 25543}.
  Type: [\code{character}].
}
\description{
In accordance with
\href{http://docs.cntd.ru/document/1200107843}{GOST 25543} (\emph{Table 5})
determine the type of brown coal by maximum \emph{moisture-holding capacity}.
}
\examples{
  brown_type(c(5.678, 19.94, 19.95, 20, 30, 69.9444))
  # [1] "10" "10" "20" "20" "30" "60"

  # building test
  stopifnot(
    all(
      brown_type(c(5.678, 19.94, 19.95, 20, 30, 69.9444)) ==
      c("10", "10", "20", "20", "30", "60")
    )
  )


}
