% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuelcat.R
\name{fuelcat}
\alias{fuelcat}
\title{Determine the category of fossil fuels in accordance with GOST 25543}
\usage{
fuelcat(sok)
}
\arguments{
\item{sok}{total volume of all \emph{fusinite} fractions, [\emph{\%}]. Type: [\code{numeric}].}
}
\value{
Identifier of fossil fuel category according to \emph{Table 4} of
  \href{http://docs.cntd.ru/document/1200107843}{GOST 25543}.
  Type: [\code{character}].
}
\description{
In accordance with
\href{http://docs.cntd.ru/document/1200107843}{GOST 25543} (\emph{Table 4})
determine the category of fossil fuel by total volume of
all \emph{fusinite} fractions. The next \emph{fusinite} fractions may be
considered:
\describe{
  \item{\emph{inertinite}}{determined in accordance with \href{https://www.iso.org/standard/42831.html}{ISO 7404-3}}
  \item{two thirds of \emph{semivitrinite}}{determined in accordance with \href{http://docs.cntd.ru/document/1200105478}{GOST R 55662}}
}
}
\examples{
  fuelcat(c(5.0, 10, 59, 80))
  # [1] "0" "1" "5" "7"
}
