% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuelclass.R
\name{fuelclass}
\alias{fuelclass}
\title{Determine the class of fossil fuels in accordance with GOST 25543}
\usage{
fuelclass(r)
}
\arguments{
\item{r}{reflectance of vitrinite, [\emph{\%}], measured in accordance with
\href{https://www.iso.org/standard/42832.html}{ISO 7404-5}.
Type: [\code{double}].}
}
\value{
Identifier of fossil fuel class according to \emph{Table 3} of
  \href{http://docs.cntd.ru/document/1200107843}{GOST 25543}.
  Type: [\code{character}].
}
\description{
In accordance with
\href{http://docs.cntd.ru/document/1200107843}{GOST 25543} (\emph{Table 3})
determine the class of fossil fuel by its reflectance of vitrinite.
}
\examples{
fuelclass(c(0.20, 0.52, 1.3, 4.99, 4.998))
# [1] "02" "05" "13" "49" "50"

}
