% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hard_subtype.R
\name{hard_subtype}
\alias{hard_subtype}
\title{Determine the subtype of hard coal in accordance with GOST 25543}
\usage{
hard_subtype(y, sl = rep_len(0, length(y)))
}
\arguments{
\item{y}{plastometric index \emph{y}, [\emph{mm}], measured in accordance with
\href{http://docs.cntd.ru/document/1200121074}{GOST 1186}.
Type: [\code{double}].}

\item{sl}{\emph{crucible swelling number}, [], measured in accordance with
\strong{ISO 501}. Type: [\code{double}].}
}
\value{
Identifier of hard coal subtype according to \emph{Table 9} of
  \href{http://docs.cntd.ru/document/1200107843}{GOST 25543}.
  Type: [\code{character}].
}
\description{
In accordance with
\href{http://docs.cntd.ru/document/1200107843}{GOST 25543} (\emph{Table 9})
determine the subtype of hard coal by plastometric index \emph{y}, [\emph{mm}],
and \emph{crucible swelling number}, [].
}
\examples{
hard_subtype(0:9, 0:9)
# [1] "00" "01" "01" "01" "01" "01" "06" "07" "08" "09"

# build test:
stopifnot(
  all(
    hard_subtype(0:9, 0:9) ==
    c("00", "01", "01", "01", "01", "01", "06", "07", "08", "09")
  )
)

}
