% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brown_code.R
\name{brown_code}
\alias{brown_code}
\title{Form the digital code for brown coals in accordance with GOST 25543}
\usage{
brown_code(r, sok, wmaxaf, tscdaf)
}
\arguments{
\item{r}{reflectance of vitrinite, [\emph{\%}], measured in accordance with
\strong{ISO 7404-5}.
Type: [\code{double}].}

\item{sok}{total volume of all \emph{fusinite} fractions, [\emph{\%}].
The next \emph{fusinite} fractions may be considered:
\describe{
  \item{\emph{inertinite}}{determined in accordance with \strong{ISO 7404-3}}
  \item{two thirds of \emph{semivitrinite}}{determined in accordance with \href{http://docs.cntd.ru/document/1200105478}{GOST R 55662}}
}
Type: [\code{numeric}].}

\item{wmaxaf}{maximum \emph{moisture-holding capacity}, [\emph{\%}], measured in
accordance with \strong{ISO 1018}. Type: [\code{double}].}

\item{tscdaf}{\emph{yield of tar}, [\emph{\%}], measured in accordance with
\strong{ISO 647} and recalculated on dry ash-free basis.
Type: [\code{double}].}
}
\value{
Digital code according to \emph{Part 7} of
  \href{http://docs.cntd.ru/document/1200107843}{GOST 25543}.
  Type: [\code{character}].
}
\description{
In accordance with
\href{http://docs.cntd.ru/document/1200107843}{GOST 25543} (\emph{part 7})
form the digital code for brown coal using results of laboratory measurements.
}
\examples{
 # Consider samples of brown coals with the next laboratory
 # measurement results:
 r0   <- c(.3)
 sok  <- c(6)
 wmaxaf <- c(53)
 tscdaf <- c(7)

 x <- brown_code(r0, sok, wmaxaf, tscdaf)
 print(x)

 # Unit test:
 #stopifnot(
 #  x \%in\% coal.state:::g25543db()$hard$S1KCFx$code
 #)
}
